+++ lib/src/tools/github.dart
@@ -18,6 +18,30 @@ class _GithubEnv {
         ? Directory(githubWorkspace)
         : Directory.current.subDir('../..');
   }
+
+  Future<void> setOutput(
+    String name,
+    Object? value, {
+    bool multiline = false,
+  }) async {
+    final githubOutput = Platform.environment['GITHUB_OUTPUT'];
+    if (githubOutput == null) {
+      throw Exception('Cannot set output! GITHUB_OUTPUT env var is not set');
+    }
+
+    final githubOutputFile = File(githubOutput);
+    if (multiline) {
+      await githubOutputFile.writeAsString(
+        '$name<<EOF\n${value}\nEOF\n',
+        mode: FileMode.append,
+      );
+    } else {
+      await githubOutputFile.writeAsString(
+        '$name=$value\n',
+        mode: FileMode.append,
+      );
+    }
+  }
 }
 
 abstract base class Github {