+++ tool/ci_gen/common/tools.dart
@@ -2,7 +2,7 @@ abstract class Tools {
   Tools._();
 
   /// https://github.com/actions/checkout/releases
-  static const actionsCheckout = 'actions/checkout@v3';
+  static const actionsCheckout = 'actions/checkout@v4';
 
   /// https://github.com/actions/upload-artifact/releases
   static const actionsUploadArtifact = 'actions/upload-artifact@v3';
@@ -20,16 +20,16 @@ abstract class Tools {
   static const dartLangSetupDart = 'dart-lang/setup-dart@v1';
 
   /// https://github.com/docker/setup-qemu-action/releases
-  static const dockerSetupQemuAction = 'docker/setup-qemu-action@v2';
+  static const dockerSetupQemuAction = 'docker/setup-qemu-action@v3';
 
   /// https://github.com/docker/setup-buildx-action/releases
-  static const dockerSetupBuildxAction = 'docker/setup-buildx-action@v2';
+  static const dockerSetupBuildxAction = 'docker/setup-buildx-action@v3';
 
   /// https://github.com/docker/login-action/releases
-  static const dockerLoginAction = 'docker/login-action@v2';
+  static const dockerLoginAction = 'docker/login-action@v3';
 
   /// https://github.com/docker/build-push-action/releases
-  static const dockerBuildAndPushAction = 'docker/build-push-action@v4';
+  static const dockerBuildAndPushAction = 'docker/build-push-action@v5';
 
   /// https://github.com/subosito/flutter-action/releases
   static const subositoFlutterAction = 'subosito/flutter-action@v2';